Get the weave and weaveDNS IPs at launch
Asking docker something (e.g., for a container's info, to get the IP
address) while a drvier operation is in progress is liable to
deadlock. This used to just be the case for `docker ps` operations,
but now seems to be the case for `docker inspect`.

To work around this somewhat disastrous problem, we do our docker
inspecting only when starting up.